fix: redirect on loading page
