+++ src/pages/donations/LoadingPage/index.tsx
@@ -5,13 +5,21 @@ import useNavigation from "hooks/useNavigation";
 import { useEffect } from "react";
 import { logEvent } from "lib/events";
 import { useCouponContext } from "contexts/couponContext";
-import { APP_INTEGRATION_LINK } from "utils/constants";
+import { APP_INTEGRATION_LINK, RIBON_COMPANY_ID } from "utils/constants";
 import extractUrlValue from "lib/extractUrlValue";
 import {
   getUTMFromLocationSearch,
   utmParamsToString,
 } from "lib/getUTMFromLocationSearch";
 import { useCurrentUser } from "contexts/currentUserContext";
+import { useCollectTickets } from "hooks/useCollectTickets";
+import { logError } from "services/crashReport";
+import { useReceiveTicketToast } from "hooks/toastHooks/useReceiveTicketToast";
+import { setLocalStorageItem } from "lib/localStorage";
+import {
+  RECEIVED_TICKET_AT_KEY,
+  RECEIVED_TICKET_FROM_INTEGRATION,
+} from "lib/localStorage/constants";
 import * as S from "./styles";
 
 function LoadingPage(): JSX.Element {
@@ -22,13 +30,9 @@ function LoadingPage(): JSX.Element {
   const externalId = extractUrlValue("external_id", history.location.search);
   const couponId = extractUrlValue("coupon_id", history.location.search);
   const { currentUser } = useCurrentUser();
-
-  useEffect(() => {
-    if (integration) {
-      localStorage.setItem("integrationName", integration.name);
-      logEvent("P1_view");
-    }
-  }, [integration]);
+  const { hasReceivedTicketToday, handleCanCollect, handleCollect } =
+    useCollectTickets();
+  const { showReceiveTicketToast } = useReceiveTicketToast();
 
   const redirectToDeeplink = () => {
     const externalIdParam = externalId ? `&external_id=${externalId}` : "";
@@ -39,37 +43,69 @@ function LoadingPage(): JSX.Element {
       `${APP_INTEGRATION_LINK}?integration_id=${integrationId}&${externalIdParam}&${couponIdParam}${utmParamsString}`,
     );
   };
-  const renderOnboardingPage = () => {
-    if (!currentUser) {
-      navigateTo({
-        pathname: "/intro",
-      });
-    } else {
-      navigateTo("/causes");
-    }
-  };
 
-  const renderReceiveCouponPage = () => {
-    if (!currentUser) {
-      navigateTo("/coupons/sign-in");
-    } else {
-      navigateTo("/coupons/give-ticket");
-    }
-  };
-
-  useEffect(() => {
-    if (process.env.REACT_APP_DEBUG_VIEW) return;
-    if (!history.location.search?.includes("_branch_match_id") && integrationId)
+  async function receiveTicket() {
+    if (
+      !history.location.search?.includes("_branch_match_id") &&
+      integrationId &&
+      !process.env.REACT_APP_DEBUG_VIEW
+    )
       redirectToDeeplink();
-  }, [integrationId]);
+    else {
+      try {
+        const isRibonIntegration =
+          integration?.id === parseInt(RIBON_COMPANY_ID, 10);
+        const canCollect = await handleCanCollect();
+        const receivedTicketToday = await hasReceivedTicketToday();
+
+        if (couponId !== "" && couponId !== undefined) {
+          setCouponId(couponId);
+          navigateTo("/coupons/give-ticket");
+        } else if (canCollect) {
+          if (currentUser && externalId) {
+            navigateTo("/intro/receive-tickets");
+          } else if (currentUser && !receivedTicketToday) {
+            if (!isRibonIntegration) {
+              navigateTo("/intro/receive-tickets");
+            } else {
+              await handleCollect({
+                onSuccess: () => {
+                  logEvent("ticketCollected", { from: "collect" });
+                },
+              });
+
+              showReceiveTicketToast();
+              setLocalStorageItem(
+                RECEIVED_TICKET_AT_KEY,
+                Date.now().toString(),
+              );
+              setLocalStorageItem(
+                RECEIVED_TICKET_FROM_INTEGRATION,
+                integrationId?.toLocaleString() ?? RIBON_COMPANY_ID,
+              );
+              navigateTo("/causes");
+            }
+          } else if (!currentUser) {
+            navigateTo("/intro");
+          } else {
+            navigateTo("/causes");
+          }
+        } else {
+          navigateTo("/causes");
+        }
+      } catch (error) {
+        logError(error);
+      }
+    }
+  }
 
   useEffect(() => {
-    if (couponId) {
-      setCouponId(couponId);
-      renderReceiveCouponPage();
+    if (integration) {
+      localStorage.setItem("integrationName", integration.name);
+      logEvent("P1_view");
     }
-    if (integration) renderOnboardingPage();
-  }, [integration, couponId]);
+    receiveTicket();
+  }, [integration]);
 
   return (
     <S.Container>
